% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georgia-data.R
\docType{data}
\name{georgia}
\alias{georgia}
\title{Georgia dataset}
\format{
A data frame with with 159 observations on the following 13 variables:
\itemize{
\item \code{AreaKey} - a numeric vector
\item \code{Latitude} - a numeric vector
\item \code{Longitud} - a numeric vector
\item \code{TotPop90} - a numeric vector
\item \code{PctRural} - a numeric vector
\item \code{PctBach} - a numeric vector
\item \code{PctEld} - a numeric vector
\item \code{PctFB} - a numeric vector
\item \code{PctPov} - a numeric vector
\item \code{PctBlack} - a numeric vector
\item \code{ID} - a numeric vector
\item \code{X} - a numeric vector
\item \code{Y} - a numeric vector
}
}
\usage{
data(georgia)
}
\description{
The Georgia census data set from Fotheringham et al. (2002) in shapefile format.
}
\keyword{datasets}
