% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mgwnbr_model}
\alias{print.mgwnbr_model}
\title{Print method for mgwnbr_model objects}
\usage{
\method{print}{mgwnbr_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mgwnbr_model}.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
The object (invisibly).
}
\description{
Print method for mgwnbr_model objects
}
