% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_print.R
\name{print.summary.mgwnbr_model}
\alias{print.summary.mgwnbr_model}
\title{Print method for summary.myclass objects}
\usage{
\method{print}{summary.mgwnbr_model}(object, ...)
}
\arguments{
\item{...}{Further arguments passed to print.}

\item{x}{An object of class \code{summary.mgwnbr_model}.}
}
\value{
The object (invisibly).
}
\description{
Print method for summary.myclass objects
}
