% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mgwnbr_model}
\alias{summary.mgwnbr_model}
\title{Summary method for mgwnbr_model objects}
\usage{
\method{summary}{mgwnbr_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mgwnbr_model}.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
An object of class \code{summary.mgwnbr_model}.
}
\description{
Summary method for mgwnbr_model objects
}
